+++ src/components/shared/modals/auth-modal/forms/login-form.tsx
@@ -12,6 +12,7 @@ import {
 	CardFooter,
 	CardHeader,
 	CardTitle,
+	Separator,
 } from '@/components/ui'
 import { FormInput } from '@/components/shared'
 import { loginUser, loginUserWithCreds } from '@/app/actions'
@@ -83,6 +84,12 @@ export const LoginForm = ({ onClose }: Props) => {
 						>
 							Увійти
 						</Button>
+
+						<Separator className='relative my-4'>
+							<span className='absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 bg-white/100 px-2 text-sm text-muted-foreground'>
+								або
+							</span>
+						</Separator>
 
 						<div className='flex w-full gap-2'>
 							<Button
