+++ src/services/api-client.ts
@@ -1,4 +1,3 @@
-import * as auth from './auth'
 import * as cart from './cart'
 import * as delivery from './delivery'
 import * as ingredients from './ingredients'
@@ -6,7 +5,6 @@ import * as products from './products'
 import * as stories from './stories'
 
 export const Api = {
-	auth,
 	cart,
 	delivery,
 	ingredients,
